obsluga zamknietej aukcji

+++ lib/bazarka_allegro/product_allegro.rb
@@ -162,9 +162,9 @@ module BazarkaAllegro
           if e.message =~ /ERR_YOU_CANT_CHANGE_ITEM/i
             # nie możemy usuwać extension dopuki nie pobierzemy wszystkich eventów
             # product.extension_for_products.where(key: 'allegro').first.destroy
-            raise "Savon::SOAPFault 1 - #{e}"
           else
-            raise "Savon::SOAPFault 2 - #{e}"
+            Rails.logger.info "Savon::SOAPFault - #{e}\n#{e.backtrace.join("\n")}"
+            raise "Savon::SOAPFault - #{e}"
           end
         rescue Exception => e
           Rails.logger.info "Exception logger - #{e}\n#{e.backtrace.join("\n")}"
@@ -172,8 +172,21 @@ module BazarkaAllegro
         end
 
       else
-        response = @allegro.do_finish_item(product.extension_for_products.where(key: 'allegro').first.allegro_id)
-        product.extension_for_products.where(key: 'allegro').first.destroy
+        begin
+          response = @allegro.do_finish_item(product.extension_for_products.where(key: 'allegro').first.allegro_id)
+        rescue Savon::SOAPFault => e
+          if e.message =~ /ERR_YOU_CANT_CHANGE_ITEM/i
+            product.extension_for_products.where(key: 'allegro').first.destroy
+          else
+            Rails.logger.info "Savon::SOAPFault - #{e}\n#{e.backtrace.join("\n")}"
+            raise "Savon::SOAPFault - #{e}"
+          end
+        rescue Exception => e
+          Rails.logger.info "Exception logger - #{e}\n#{e.backtrace.join("\n")}"
+          raise "Exception - #{e}"
+        ensure
+          product.extension_for_products.where(key: 'allegro').first.destroy
+        end
       end
       #Rails.logger.info response.inspect
       response
